+++ Runtime/Buttons/DelayedButton.cs
@@ -11,13 +11,14 @@ namespace ActionCode.UI
     /// <para>It will trigger the button On Click action after the <see cref="delay"/>.</para>
     /// </summary>
     [AddComponentMenu("UI/Delayed Button")]
-    public sealed class DelayedButton : Button, ISelectable, ISubmitable
+    public sealed class DelayedButton : Button, ISelectable, ISubmitable, IHighlightable
     {
         [SerializeField, Tooltip("Time (in seconds) to trigger the button On Click action."), Min(minDelay)]
         private float delay = 0.2F;
 
         public event Action OnSelected;
         public event Action OnSubmitted;
+        public event Action<GameObject> OnHighlighted;
 
         /// <summary>
         /// Time (in seconds) to trigger the button On Click action.
@@ -43,6 +44,12 @@ namespace ActionCode.UI
         {
             var isLeftClick = eventData.button == PointerEventData.InputButton.Left;
             if (isLeftClick) TrySubmit(eventData);
+        }
+
+        public override void OnPointerEnter(PointerEventData eventData)
+        {
+            base.OnPointerEnter(eventData);
+            OnHighlighted?.Invoke(gameObject);
         }
 
         private void TrySubmit(BaseEventData eventData)